Sort database before saving.
Make CblPkg an instance of Ord so that a CblDB ([CblPkg]) can be sorted.  Do
that sort when saving a database, which should have the effect that databases
are easier to compare, and also make it easier to keep a database under
version control.

The implementation of the Ord instance for CblPkg is based on sorting of
tuples.

Signed-off-by: Magnus Therning <magnus@therning.org>