Bump of applicative and aeson.
The bump of applicativs to 0.11 required quite a few changes due to an API
change.  It also high-lighted that the command-line-parser helpers in
Util.Misc aren't very well-written.  Making even a minimal adaptation to the
new API required a complete rewrite of one of the parsers.

I've chosen to use ReadP to write the parsers.  It seems like a lighter option
than bringing in Parsec, while still being comparatively expressive.

Signed-off-by: Magnus Therning <magnus@therning.org>