Remove auto-derived {From,To}JSON instances for Dependency.
The PackageName type changed slightly between 1.18 and 1.22, which impacted
the JSON instances and made existing databases unreadable.  This change
switches to a fixed JSON mapping for Dependency.  Arguably the same should be
done for all other non-local types.

The change in OldPkgDB.hs is untested, but I'm certain a change is needed and
fairly certain this one is good.

Signed-off-by: Magnus Therning <magnus@therning.org>