Don't choke on index entries for preferred versions.
The index now contains files listing preferred versions of packages, e.g.
`aeson/preferred-versions`.  Unfortunately that broke an assumption on the
format of the contents of the index: not everything maches
`<pkg-name>/<version>/<cbl-file>`!

With this change an attempt is made to parse the second part of the path as a
version, and if that failes (which it does for `preferred-versions`) the file
is skipped.

Signed-off-by: Magnus Therning <magnus@therning.org>